carbon: Made the vertical table header a client component
