+++ carbon/components/charts/helpers/DataTable/PaginatedTable/index.tsx
@@ -1,11 +1,11 @@
+import Skeleton from "components/Skeleton";
 import {
   ConfigurationKey,
   fetchData,
 } from "components/charts/helpers/DataTable/configurations";
-import Table from "components/charts/helpers/DataTable/Table";
-import Skeleton from "components/Skeleton";
-import { PaginatedResponse, SortQueryParams } from "lib/charts/types";
-import { Dispatch, SetStateAction, useEffect, useState } from "react";
+import { PaginatedResponse } from "lib/charts/types";
+import { useEffect, useState } from "react";
+import TableHeaderClientWrapper from "../TableHeaderClientWrapper";
 
 /** A Client Component (wannabe Server Component) that renders the actual Table
  * configurationKey: Table configuration key
@@ -17,8 +17,6 @@ export default function PaginatedTable<RI>(props: {
   page: number;
   params: object;
   skeletonClassName?: string;
-  sortParams: SortQueryParams
-  setSortParams: Dispatch<SetStateAction<SortQueryParams>>
 }) {
   const [data, setData] = useState<PaginatedResponse<RI> | null>(null);
   useEffect(() => {
@@ -30,13 +28,11 @@ export default function PaginatedTable<RI>(props: {
   }, [props.page]);
 
   return data ? (
-    <Table
+    <TableHeaderClientWrapper
       configurationKey={props.configurationKey}
       params={props.params}
       data={data}
-      sortParams={props.sortParams}
-      setSortParams={props.setSortParams}
-    ></Table>
+    />
   ) : (
     <Skeleton className={props.skeletonClassName} />
   );
