+++ carbon/components/charts/helpers/DataTable/TableHeaderClientWrapper/styles.module.scss
@@ -1,0 +1,19 @@
+@use "~@klimadao/lib/theme/breakpoints.scss";
+.table {
+  color: var(--lightmode-font-01);
+}
+.desktopOnly {
+  display: none;
+
+  @include breakpoints.medium {
+    display: block;
+  }
+}
+
+.mobileOnly {
+  display: block;
+
+  @include breakpoints.medium {
+    display: none;
+  }
+}
