+++ carbon/components/charts/helpers/DataTable/configurations/types.ts
@@ -1,4 +1,4 @@
-import { PaginatedResponse, SortQueryParams } from "lib/charts/types";
+import { PaginatedResponse } from "lib/charts/types";
 import React from "react";
 
 type Formatter<RI> =
@@ -24,8 +24,6 @@ export interface Columns<RI> {
 export interface DataRendererProps<RI, P> {
   data: PaginatedResponse<RI>;
   params?: P;
-  sortParams: SortQueryParams;
-  setSortParams: (sortParams: SortQueryParams) => void;
 }
 export type DataRenderer<RI, P> = (
   props: DataRendererProps<RI, P>
